% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_est.R
\name{adapt_est}
\alias{adapt_est}
\title{Estimate inverse probability weighted mean outcome for each arm}
\usage{
adapt_est(adapt_matrix = NULL)
}
\arguments{
\item{adapt_matrix}{(default = \code{NULL}) result of \code{run_adapt()}}
}
\value{
\code{adapt_est} returns a list containing the following elements:
 \itemize{
   \item \code{est}: IPW estimates in an object of \code{estimatr::lm_robust()}.
   \item \code{est_plot}: coefficient plot created with \code{ggplot2::ggplot()}.
 }
}
\description{
Estimates mean outcome using inverse probability weight and creates a coefficient plot based on the result
}
\references{
Lee and Green. (2023+). Discovering optimal ballot wording using adaptive survey design.
}
