% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_survey_data}
\alias{ex_survey_data}
\title{Example survey data to be used for demonstration}
\format{
data
}
\description{
Example survey data to be used for demonstration
}
\keyword{datasets}
