% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Creates a plot for posterior probability}
\usage{
plot_posterior(adapt_matrix = NULL)
}
\arguments{
\item{adapt_matrix}{(default = \code{NULL}) result of \code{run_adapt()}}
}
\value{
\code{plot_posterior} returns ggplot-based plot
}
\description{
Creates a ggplot showing over time development of the posterior probability that each arm is best from an adaptive experiment.
}
\references{
Lee and Green. (2023+). Discovering optimal ballot wording using adaptive survey design.
}
