% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_adapt.R
\name{run_adapt}
\alias{run_adapt}
\title{Conducts simple adaptive survey with standard Bernoulli Thompson sampling}
\usage{
run_adapt(
  prior_survey_data = NULL,
  prior_adapt_matrix = NULL,
  current_period = NULL,
  arms = NULL,
  periods = NULL,
  floor = NULL,
  seed = 1004
)
}
\arguments{
\item{prior_survey_data}{(default = \code{NULL}) data.frame containing survey results from prior period. Must contain the following two columns: \code{arm} that contains treatment arm numbers 1 through k, and \code{Y} that contains binary survey outcome (0 or 1). Ignored in the first period (i.e., when \code{current_period == 1}).}

\item{prior_adapt_matrix}{(default = \code{NULL}) result of \code{run_adapt()} from the previous batch. Ignored in the first period (i.e., when \code{current_period == 1}).}

\item{current_period}{(default = \code{NULL}) numeric value indicating the number of current batch to be implemented (i.e., total number of batches previously done + 1). For example, if a user has conducted a survey for the first batch and is looking to update treatment probability for the second batch, the value should be 2.}

\item{arms}{(default = \code{NULL}) Total number of arms in the survey.}

\item{periods}{(default = \code{NULL}) Total number of periods to be implemented in the survey.}

\item{floor}{(default = \code{NULL}) A numeric value between 0 and 1 indicating the minimum probability of treatment assignment imposed on each arm.}

\item{seed}{(default = 1004) Integer to specify seeds.}
}
\value{
\code{run_adapt} returns a list containing the following elements:
 \itemize{
   \item \code{m_posterior}: matrix containing posterior probability for each arm (columns) for each period (rows).
   \item \code{m_trials}: matrix containing cumulative number of observations assigned to each arm (columns) for each period (rows).
   \item \code{m_success}: matrix containing cumulative number of successes (i.e., when outcome equals 1) for to each arm (columns) for each period (rows).
   \item \code{prior_survey_data}: appended survey data with batch number, treatment assignment probability, and inverse-probability weight appended.
 }
}
\description{
Allows users to iteratively update treatment assignment probability in an adaptive survey.
Returns all necessary statistics including success rate and posterior probabilities.
This function is to be used iteratively in an adaptive design, as the elements in a returned list will be iteratively updated over periods.
}
\references{
Lee and Green. (2023+). Discovering optimal ballot wording using adaptive survey design.
}
