% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate adaptive or static experiments while varying several parameters and produce key simulated statistics}
\usage{
simulate(
  probs = NA,
  static = FALSE,
  periods = 10,
  n = 1000,
  n_first = NA,
  iter = 1000,
  floor_rate = 0.01
)
}
\arguments{
\item{probs}{(default = \code{NULL}) distribution of true mean outcomes for each arm}

\item{static}{(default = \code{FALSE}) whether the simulated experiment should be adaptive or static}

\item{periods}{(default = \code{10}) total number of batches in a single experiment (relevant for adaptive design only)}

\item{n}{(default = \code{1000}) total sample size}

\item{n_first}{(default = \code{NULL}) total sample size to be allotted to the first batch. If \code{NULL}, \code{n} is evenly distributed across batches}

\item{iter}{(default = \code{1000}) number of simulations to run}

\item{floor_rate}{(default = \code{0.01}) minimum sample probability to implement to ensure all arms receive non-zero sample in each batch}
}
\value{
\code{simulate} returns a list containing the following elements:
 \itemize{
   \item \code{outmat}: simulation summary statistics containing proportion of iterations that each arm is identified as the best arm, bias, RMSE, and coverage.
   \item \code{modmat}: IPW estimates (result of \code{estimatr::lm_robust()}) for each arm for all iterations.
   \item \code{postmat}: data.frame containing posterior probabilities for each arm (columns) for each period (rows) for all iterations (rows).
 }
}
\description{
Simulate experiments varying arms, periods, sample size, and design and outputs simulated results including selection of best arms, RMSE, bias, coverage.
}
\references{
Lee and Green. (2023+). Discovering optimal ballot wording using adaptive survey design.
}
